// components/student/studentTestResult/studentTestResultQuestion/studentTestResultQuestionStyle.scss.d.ts
export type Styles = {
  'animationShowQuestionResultAnswers': string;
  'answersContainer': string;
  'answersContainerEmpty': string;
  'buttonShowAnswers': string;
  'buttonShowAnswersCloseFalse': string;
  'buttonShowAnswersCloseTrue': string;
  'buttonShowAnswersContainer': string;
  'buttonShowAnswersFalse': string;
  'buttonShowAnswersOpenFalse': string;
  'buttonShowAnswersOpenTrue': string;
  'buttonShowAnswersTrue': string;
  'componentContainer': string;
  'numberOfPointsContainer': string;
  'questionContent': string;
  'questionContentContainer': string;
  'questionFalse': string;
  'questionTrue': string;
};

export type ClassNames = keyof Styles;

declare const styles: Styles;

export default styles;